+++ lib/screens/roles/switch_roles.dart
@@ -5,7 +5,7 @@ import 'package:provider/provider.dart';
 import '../../core/models/user.dart';
 import '../../core/routes/router.dart';
 import '../../core/routes/routes.dart';
-import '../../router/roles.dart';
+import '../../core/router/roles.dart';
 import '../../styles/icons/chap_chap_icons.dart';
 import '../../styles/ui/colors.dart';
