+++ lib/screens/shared/about.dart
@@ -1,4 +1,4 @@
-import 'package:client/router/roles.dart';
+import 'package:client/core/router/roles.dart';
 import 'package:client/styles/icons/chap_chap_icons.dart';
 import 'package:flutter/material.dart';
