+++ lib/screens/shared/profile.dart
@@ -1,4 +1,4 @@
-import 'package:client/router/roles.dart';
+import 'package:client/core/router/roles.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
