+++ lib/screens/shared/settings.dart
@@ -9,7 +9,7 @@ import 'package:location_picker_flutter_map/location_picker_flutter_map.dart';
 import 'package:provider/provider.dart';
 
 import '../../core/providers/user.dart';
-import '../../router/roles.dart';
+import '../../core/router/roles.dart';
 import '../../styles/icons/chap_chap_icons.dart';
 import '../../styles/ui/colors.dart';
 import '../auth/login.dart';
